requirements.txt
Flask
google-generativeai
pywebview